Add check table exists

+++ database/migrations/2017_08_11_074006_create_menu_items_wp_table.php
@@ -13,23 +13,25 @@ class CreateMenuItemsWpTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('menu.table_prefix') . config('menu.table_name_items'), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('label');
-            $table->string('link');
-            $table->string('icon')->nullable();
-            $table->unsignedBigInteger('parent')->default(0);
-            $table->integer('sort')->default(0);
-            $table->string('class')->nullable();
-            $table->enum('target', ['_self', '_blank'])->default('_self');
-            $table->unsignedBigInteger('menu');
-            $table->integer('depth')->default(0);
-            $table->timestamps();
+        if (!Schema::hasTable(config('menu.table_prefix') . config('menu.table_name_items'))) {
+            Schema::create(config('menu.table_prefix') . config('menu.table_name_items'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('label');
+                $table->string('link');
+                $table->string('icon')->nullable();
+                $table->unsignedBigInteger('parent')->default(0);
+                $table->integer('sort')->default(0);
+                $table->string('class')->nullable();
+                $table->enum('target', ['_self', '_blank'])->default('_self');
+                $table->unsignedBigInteger('menu');
+                $table->integer('depth')->default(0);
+                $table->timestamps();
 
-            $table->foreign('menu')->references('id')->on(config('menu.table_prefix') . config('menu.table_name_menus'))
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('menu')->references('id')->on(config('menu.table_prefix') . config('menu.table_name_menus'))
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
